requirements.txt
Django~=4.2.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
Pillow
